AP_TECS: fix calculation of TAS rate demand.
The effects of this but where especially visible when flying in AUTO modes
with constantly changing values of speed demand (e.g. with throttle nudge,
or ground speed undershoot enabled).
In the old case (_TAS_rate_dem = (_TAS_dem - _TAS_dem_last) / 0.1f;) consider
the scenarios where TAS_dem was way out of the (velRateMin,velRateMax) limit
(so previous iteration had run with limiting) and then we jump back to the valid range.
In this case the difference between _TAS_dem_last and _TAS_dem would be large
and not limited in any manner, which in turn created a single nasty value of _TAS_rate_dem.